Safari-609.1.11.1/LayoutTests/platform/ios/ios/fast/css/webkit-overflow-scrolling-parsing-expected.txt

This tests checks parsing of the '-webkit-overflow-scrolling' property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS test("-webkit-overflow-scrolling: auto", "-webkit-overflow-scrolling") is "auto"
PASS test("-webkit-overflow-scrolling: banana", "-webkit-overflow-scrolling") is null
PASS test("-webkit-overflow-scrolling: touch", "-webkit-overflow-scrolling") is "touch"
PASS successfullyParsed is true

TEST COMPLETE
